Move rest clients to packages
