+++ UPGRADE.md
@@ -1,5 +1,21 @@
 # Upgrade to 2.0
 
+## Default REST clients moved to external packages
+
+`simpleRestClient` and `jsonServerRestClient` are no longer part of the core package. They have been moved to standalone packages, where they are the default export:
+
+* `simpleRestClient` => `ra-data-simple-rest`
+* `jsonServerRestClient` => `ra-data-json-server`
+
+Update your `import` statements accordingly:
+
+```js
+// before
+import { simpleRestClient } from 'admin-on-rest';
+// after
+import simpleRestClient from 'ra-data-simple-rest';
+```
+
 ## `addField` Prop Replaced By `addField` HOC
 
 Adding the `addField` prop to a component used to automatically add a redux-form `<Field>` component around an input component that you wanted to bind to the edit or create form. This feature was moved to a Higher-order component (HOC):
