+++ packages/ra-data-json-server/package.json
@@ -1,0 +1,25 @@
+{
+    "name": "ra-data-json-server",
+    "version": "2.0.0-alpha1",
+    "description": "JSON Server data provider for react-admin",
+    "files": [
+        "*.md",
+        "lib",
+        "src"
+    ],
+    "main": "index",
+    "scripts": {
+        "test": "make test"
+    },
+    "authors": [
+        "François Zaninotto"
+    ],
+    "repository": "marmelab/react-admin",
+    "homepage": "https://github.com/marmelab/react-admin#readme",
+    "bugs": "https://github.com/marmelab/react-admin/issues",
+    "license": "MIT",
+    "dependencies": {
+        "react-admin": "2.0.0-alpha1",
+        "query-string": "~5.0.0"
+    }
+}
