+++ packages/react-admin/index.js
@@ -1,3 +1,4 @@
+export * from './dataFetchActions';
 export * from './actions';
 export * from './auth';
 export * from './i18n';
@@ -6,7 +7,6 @@ export createAppReducer from './reducer';
 export adminReducer from './reducer/admin';
 export localeReducer from './reducer/locale';
 export queryReducer from './reducer/admin/resource/list/queryReducer';
-export * from './rest';
 export * from './sideEffect/saga';
 export * as fetchUtils from './util/fetch';
 export FieldTitle from './util/FieldTitle';
