--- packages/react-admin/rest/index.js
@@ -1,1 +1,0 @@
-export * from './types';
